Move command implementation to `commands.ts`
